% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMfunctions.R
\name{Adagrad}
\alias{Adagrad}
\title{Adagrad
Implementation of the Adagrad algorithm. Left TBD at the moment.}
\usage{
Adagrad(wInitVals, data, labels)
}
\description{
Adagrad
Implementation of the Adagrad algorithm. Left TBD at the moment.
}
