% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMfunctions.R
\name{ADAM}
\alias{ADAM}
\title{ADAM
Implementation of the ADAM algorithm. Left TBD at the moment.}
\usage{
ADAM(wInitVals, data, labels)
}
\description{
ADAM
Implementation of the ADAM algorithm. Left TBD at the moment.
}
