% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AUC}
\alias{AUC}
\title{Area Under Curve}
\usage{
AUC(TPR, FPR)
}
\arguments{
\item{TPR}{is a vector of true positive rates.}

\item{FPR}{is a vector of the false positive rates.}
}
\description{
Calculates the area under the curve.
}
\examples{



}
\keyword{AUC}
