% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{binaryLDA}
\alias{binaryLDA}
\title{Binary LDA}
\usage{
binaryLDA(x1, x2, priors, test, classifications = c("2", "1"))
}
\arguments{
\item{x1}{contains all rows of the first class sent in of the training set}

\item{x2}{contains all rows of the first class sent in of the training set}

\item{priors}{are the priors for the 2 classes}

\item{test}{is the test set to evaluate on.}

\item{classifications}{are the factor names, sent in as strings}
}
\description{
Does an LDA on a dataset. Returns the classifications of the test set
}
\examples{


}
\keyword{LDA,}
\keyword{binary}
