% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMfunctions.R
\name{cDivT}
\alias{cDivT}
\title{cLogT}
\usage{
cDivT(t, c = 1, power = 1)

cDivT(t, c = 1, power = 1)
}
\arguments{
\item{t}{is the timestep}

\item{c}{is a coefficient to tune with.}

\item{power}{is to what power one wants to elevate t to.}

\item{t}{is the timestep}

\item{c}{is a coefficient to tune with.}

\item{power}{is to what power one wants to elevate t to.}
}
\description{
Step size function, deciding the learning rate.
Step size is set to 1/log(t), which is a common stepsize, possible to alter through c and power
Return c/(power*log(t)), with default c = 1 and power = 1

Step size function, deciding the learning rate.
Step size is set to 1/t, which is a common stepsize, possible to alter through c and power
Return c/(t^power), with default c = 1 and power = 1
}
\examples{






}
\keyword{Stepsize,}
\keyword{learning}
\keyword{rate}
