% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{classificationRate}
\alias{classificationRate}
\title{Classification rate}
\usage{
classificationRate(confMatrix)
}
\arguments{
\item{confMatrix}{is the confusion matrix to calculate the classification rate for.}
}
\description{
Calculates, out of a confusion matrix, the amount of correctly classified data points.
}
\examples{



}
\keyword{accuracy}
\keyword{confusion}
\keyword{matrix,}
