% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CVtrainAndValidationSet}
\alias{CVtrainAndValidationSet}
\title{CVtrainAndValidationSet}
\usage{
CVtrainAndValidationSet(splits, splitIndex = -1)
}
\arguments{
\item{splits}{List of partitions of the dataset. Obtained in correct format through function crossValidationSplits.}

\item{splitIndex}{Instructs which partition to take. If it is -1, it will just take a random split. Otherwise, it will take the corresponding index in the list.}

\item{stdDev}{to standardize the vector with. If nothing is sent in, the sd of the feature vector will be used.}
}
\description{
Returns a training set (k-1 folds) and a validation set (1 fold) in a list generated by the function \code{crossValidationSplits}.
}
\examples{



}
\keyword{Cross}
\keyword{Validation}
