% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{DiscFunc}
\alias{DiscFunc}
\title{DiscFunc}
\usage{
DiscFunc(x, cov_k, mu_k, prior)
}
\arguments{
\item{x}{is the data point to be scored.}

\item{cov_k}{is the cov matrix for class k. Is same for all classes in LDA}

\item{mu_k}{is the mean for class k.}

\item{prior}{is the prior for class k}
}
\description{
Discriminant function used for LDA.
}
\examples{


}
\keyword{Discriminant}
\keyword{LDA,}
\keyword{binary,}
\keyword{function}
