% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dist_plot}
\alias{dist_plot}
\title{Distribution plot}
\usage{
dist_plot(x, ...)
}
\arguments{
\item{x}{is the vector of data to plot.}

\item{...}{are additional parameters to the \code{hist} function.}
}
\description{
Plots the distribution of data.
}
