% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{fitted.GaussianNB}
\alias{fitted.GaussianNB}
\title{Fitted values GaussianNB}
\usage{
\method{fitted}{GaussianNB}(modelObj, ret_probabilities = F)
}
\arguments{
\item{modelObj}{is the model itself, the \code{GaussianNaiveBayes} model.}
}
\description{
Returns the fitted values for a Gaussian Naive Bayes model.
}
