% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{fitted.NB}
\alias{fitted.NB}
\title{Fitted Naive Bayes}
\usage{
\method{fitted}{NB}(model, ret_probabilities = F)
}
\arguments{
\item{model}{is the model in question to return the predictions for.}

\item{ret_probabilities}{is a boolean deciding whether to return probabilities or class predictions. Default \code{FALSE}.}
}
\description{
Returns the fitted values of a Naive Bayes model.
}
