% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{gauss_inf_NB}
\alias{gauss_inf_NB}
\title{Gaussian inference Naive Bayes}
\usage{
gauss_inf_NB(X, i, classes)
}
\arguments{
\item{X}{is the data matrix to predict for.}

\item{i}{is the column index.}

\item{classes}{are the response variable classes.}
}
\description{
Internal function to calculate names.
}
\keyword{internal}
