% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{GaussianNaiveBayes}
\alias{GaussianNaiveBayes}
\title{Gaussian Naive Bayes}
\usage{
GaussianNaiveBayes(X, y, priors = "weighted")
}
\arguments{
\item{X}{is the vector of features.}

\item{y}{is the response variable.}

\item{priors}{are priors used.}
}
\description{
Implement a Gaussian Naive Bayes.
}
\examples{
library(gclus)
data("wine")
wine
GNB = GaussianNaiveBayes(wine[,-c(1)], wine$Class)
table(wine$Class, fitted(GNB))
t(fitted(GNB, ret_probabilities = T))
}
