% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{Generate2DimSampleData}
\alias{Generate2DimSampleData}
\title{Generate 2-dim Sample-data from polynomial}
\usage{
Generate2DimSampleData(n, constants, xMin, xMax, var, integers = FALSE,
  seed = 123)
}
\arguments{
\item{xMin}{is the minimum x-value.}

\item{xMax}{is the maximum x-value.}

\item{var}{is the variance of the noise.}

\item{integers}{specifies whether only to generate integers or not. Default \code{FALSE}.}

\item{seed}{for reproducability. Default 123.}

\item{Constants}{is a vector with (degree + 1) constans, i.e. a + bx + cx^2 for degree = 2}

\item{xSequence}{is the x interval on which the dots are generated; all dots included.}
}
\description{
Generates 2-dimensional according to a polynomial
}
