% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_formula}
\alias{get_formula}
\title{Get formula}
\usage{
get_formula(response, feats)
}
\arguments{
\item{response}{is the name of the}

\item{feats}{is a vector containing the names of the features}
}
\description{
Returns the formula as a string
}
\examples{
get_formula("Fertility", c("Infant.Mortality", "No.In.Household"))
# Will print out "Fertility ~ Infant.Mortality + No.In.Household"
}
\keyword{error}
