% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_skewed_vars}
\alias{get_skewed_vars}
\title{Get skewness}
\usage{
get_skewed_vars(data, skewed_threshold = 1)
}
\arguments{
\item{data}{is the data frame for which to obtain the skews.}

\item{skewed_threshold}{is the absolute threshold value. Default is 1, meaning all variable having \code{abs(skewness(variable)) > 1} will be returned.}
}
\description{
Get the skewness of all variables in a data set.
}
