% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{holdout_method}
\alias{holdout_method}
\title{Holdout method}
\usage{
holdout_method(data, train_partition = 0.75, seed = NULL)
}
\arguments{
\item{data}{is the data to split}

\item{train_partition}{is the percentual size of the data set. 0.75 means 75 % of the dataset will be trained on.}

\item{seed}{is the seed to set. Default set to NULL}
}
\description{
Returns a training and a test set.
}
\keyword{error}
