% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{k_fold_cv}
\alias{k_fold_cv}
\title{K-fold cross-validation}
\usage{
k_fold_cv(model, data, error_measure_function, formula_ = NULL,
  target = NULL, predictors = NULL, fold_indices = NULL,
  n_folds = 5, seed = 123, ...)
}
\arguments{
\item{model}{is the model function sent in onto which the cross-validation should be performed.}

\item{data}{is the data given to split up}

\item{error_measure_function}{is the error function with which one measures the generalization error.}

\item{formula_}{is the formula to pass to the model function. \code{NULL} by default, and if it is \code{NULL}, \code{target} and \code{predictors} must be passed.}

\item{target}{is the name of the target variable. Is NULL by default under assumption that formula exists.}

\item{predictors}{is a vector of the names of the predictors. Is NULL by default under assumption that formula exists.}

\item{fold_indices}{is a list of fold indices. Default is \code{NULL}, meaning that the partitioning will be randomized into \code{n_folds}.}

\item{n_folds}{is the number of folds to divide.}

\item{...}{are additional model parameters.}
}
\description{
Performs cross-validation on an arbitrary model in an arbitrary way.
}
\details{
By default, if \code{fold_indices = NULL}, k-fold-CV is performed.

Not tested yet.
}
