% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMfunctions.R
\name{KernelizedPerceptron}
\alias{KernelizedPerceptron}
\title{Kernelized Perceptron}
\usage{
KernelizedPerceptron(X, y_vec, X_test, Y_test, kernel, ...)
}
\arguments{
\item{X}{is the training set features}

\item{y_vec}{is the labels for training set}

\item{X_test}{is the test set}

\item{Y_test}{is the is the labels for the training set}

\item{kernel}{is the kernel function used. The "..." are the extra parameters necessary for tuning the kernel (e.g. the variance for a gaussian kernel).}
}
\description{
This function is a very simple implementation of a kernelized perceptron
Step size is set to 1/t, which is a common stepsize, but one should be able to alter it. This is TBD
Currently WAY too slow. No point in even using unless improvements are made.
}
\examples{


}
\keyword{Kernel}
