% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeans.R
\name{KMeansOnline}
\alias{KMeansOnline}
\title{Online k-means}
\usage{
KMeansOnline(k, features, seed = 123, learnRateThreshold = 10^(-6),
  learnRateFunc = cDivT, ...)
}
\arguments{
\item{k}{is the number of clusters}

\item{features}{is the data sent in. Send in as a data frame or matrix, but ensure to send in data points as rows and features as columns.}

\item{seed}{used to set the seed for initializing. Default set to 123.}

\item{learnRateThreshold}{is the threshold of learning rate when to stop.}

\item{learnRateFunc}{is the function calculating the learning rate. Set to function cDivT (check package)}
}
\description{
Implementation of K-means online algorithm. Although it does not receive point by point, it follows the same principle as k-means online. However, not really the best algorithm perhaps. Everything depends on the learning rate.
}
\examples{



}
\keyword{cluster,}
\keyword{k-means}
