% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNearest.R
\name{knearest}
\alias{knearest}
\title{K-nearest neighbours classification}
\usage{
knearest(train, K, test, labelsTrain, labelsTest, pLimit = 0.5)
}
\arguments{
\item{train}{is the training set}

\item{K}{is the k in k-nearest neighbours. That is, the algorithm will consider the k nearest neighbours to the point in each iteration.}

\item{test}{is the test set}

\item{labelsTrain}{are the labels of the training set. Not really used at the moment in the function. TBD.}

\item{labelsTest}{are the labels of the test set.}
}
\description{
Returns the classifications of the test set.
}
\examples{


}
\keyword{Volume,}
\keyword{sphere}
