% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNearest.R
\name{kNearestNeighDensEstimation}
\alias{kNearestNeighDensEstimation}
\title{k-nearest density estimation}
\usage{
kNearestNeighDensEstimation(data, k, newData)
}
\arguments{
\item{data}{is the data given.}

\item{newData}{is the data to perform the density estimation over. Can in two dimensions be just a sequence of values (for example)}
}
\description{
Returns a density estimation over the grid of new data.
}
\examples{


}
\keyword{density}
\keyword{estimation}
\keyword{k-nearest,}
