% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{leave_one_out_cv}
\alias{leave_one_out_cv}
\title{Leave one out Cross-Validation}
\usage{
leave_one_out_cv(model, data, formula_, target = NULL,
  predictors = NULL, error_measure_function, ...)
}
\arguments{
\item{model}{is the model on which to perform the time-series cross-validation.}

\item{data}{is the data on which to perform the cross-validation and provide good measures for.}

\item{formula_}{is the formula to pass to the model function. \code{NULL} by default, and if it is \code{NULL}, \code{target} and \code{predictors} must be passed.}

\item{target}{is the name of the target variable. Can be \code{NULL} if \code{formula_} is specified.}

\item{predictors}{is a vector of the names of the predictors. Can be \code{NULL} if \code{formula_} is specified.}

\item{error_measure_function}{is the error function with which one measures the generalization error.}

\item{...}{are additional model parameters.}
}
\description{
Performs Leave-One-Out Cross-validation on the dataset given.
}
\details{
Not tested yet.
}
