% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{linear_reg}
\alias{linear_reg}
\title{Linear regression}
\usage{
linear_reg(X, y, transform_func = NULL, ...)
}
\arguments{
\item{X}{is the data frame with data.}

\item{y}{are the response variables.}

\item{transform_func}{is an optional parameter to transform the parameters into another dimensional space. Can be \code{log}, creating \deqn{y[i] = \beta[0]\prod \beta[j]x[i,j]}
for example. Default is \code{NULL}.}

\item{...}{are additional parameters to \code{transform_func}.

Note that the standard errors are not really applicable if \code{transform_func} is used.}
}
\description{
Performs linear regression in closed form.
}
