% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMeans.R
\name{LLoydsAlgorithm}
\alias{LLoydsAlgorithm}
\title{Lloyd's algorithm}
\usage{
LLoydsAlgorithm(k, features, seed = 123,
  convergenceCheckType = "No Change", threshold = 10^(-10))
}
\arguments{
\item{k}{is the number of clusters}

\item{features}{is the data sent in. Send in as a data frame or matrix, but ensure to send in data points as rows and features as columns.}

\item{seed}{used to set the seed for initializing. Default set to 123.}

\item{convergenceCheckType}{can be "No Change" or "Threshold". If none of these, it will never converge. If "Threshold", algorithm will stop when maximum change of any mu in any direction is lower than the threshold. If "No Change", the algorithm will stop when no change of assignment of points to mus has been done during an iteration.}

\item{threshold}{used if convergenceCheckType = "Threshold". Algorithm will stop when maximum change of any mu in any direction is lower than the threshold.}
}
\description{
Implementation of Lloyd's algorithm. Solves the k-means problem.
}
\examples{



}
\keyword{cluster,}
\keyword{k-means}
