% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{loss_binary_classification}
\alias{loss_binary_classification}
\title{Classification with loss}
\usage{
loss_binary_classification(FP_punishment, FN_punishment, probabilities)
}
\arguments{
\item{FP_punishment}{is the value in the loss function for all the false positives yielded.}

\item{FP_punishment}{is the value in the loss function for all the false negatives yielded.}

\item{probabilites}{is a vector or matrix with the probabilities. Can either only be the probabilites for a positive result, or a Nx2 matrix with the probability for a positive classification in the second column.}
}
\description{
Function used for binary classification to incorporate loss.
}
\keyword{binary,}
\keyword{classification}
\keyword{loss,}
\keyword{matrix,}
