% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{MAPE}
\alias{MAPE}
\title{MAPE}
\usage{
MAPE(y, y_hat)
}
\arguments{
\item{y}{is a vector of the actual values.}

\item{y_hat}{is a vector of the predictions.}
}
\description{
Calculates the MAPE value for predictions.
}
\details{
NOT tested yet!
}
