% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MSE}
\alias{MSE}
\title{MSE - Mean Squared Error}
\usage{
MSE(y, y_hat)
}
\arguments{
\item{y}{is a vector containing the actual response values.}

\item{y_hat}{is a vector containing the predictions.}
}
\description{
Calculates the mean squared error for predictions.
}
\examples{



}
\keyword{error}
