% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LDA.R
\name{multiclassLDA}
\alias{multiclassLDA}
\title{Multiclass LDA}
\usage{
multiclassLDA(train, labelsTrain, priors, validation)
}
\arguments{
\item{train}{is a dataframe containing the features we are interested to try out for. Do not send in the labels in this dataframe.}

\item{labelsTrain}{is a vector containing the labels for the training set}

\item{priors}{are the priors for the 2 classes}

\item{validation}{is the test set to evaluate on. Do not send in the labels in this dataframe.}

\item{classifications}{are the factor names, sent in as strings}
}
\description{
Does an LDA on a dataset with several different classes. Returns the classifications of the test set. Manages to do the exact same thing as lda() from MASS, except it already predicts. Should perhaps be modified to return more information such as decision boundaries, covM matrix etc, but left as TBD.
}
\examples{

}
\keyword{LDA,}
\keyword{multiclass}
