% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{multinomial_table}
\alias{multinomial_table}
\title{Inference Multinomial Naive Bayes}
\usage{
multinomial_table(X, i, classes)
}
\arguments{
\item{X}{is the data matrix to predict for.}

\item{i}{is the column index.}

\item{classes}{are the response variable classes.}
}
\description{
Internal function to calculate names.
}
\keyword{internal}
