% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{MultinomialNB}
\alias{MultinomialNB}
\title{Multinomial Naive Bayes}
\usage{
MultinomialNB(X, y, priors = "weighted")
}
\arguments{
\item{X}{should be a data frame. Should only contain categorical features (factors preferrably) the will be used for inference.}

\item{y}{are the labels.}

\item{priors}{are the priors to use for the classes. Default is "weighted", meaning the proportions of each class in the training set.}
}
\description{
This function creates an implementation of NB. The equation reads \deqn{p(y[i] == 1|x[i]) = (\Pi[j] p(y|x[i,j]))*p(y[i] == 1)}
}
\examples{
library("bnlearn")
data(asia)
asia
MNB = MultinomialNB(X = asia[,1:7], y = asia[,8])
predict(MNB, asia[,1:7])
t(predict(MNB, asia[,1:7], ret_probabilities = T))
}
\keyword{Bayes,}
\keyword{NB}
\keyword{Naive,}
