% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{NaiveBayesNonParam}
\alias{NaiveBayesNonParam}
\title{Naive Bayes non-parametric}
\usage{
NaiveBayesNonParam(train, trainLabels, test, testLabels, priors)
}
\arguments{
\item{train}{is the training data to perform the density estimation on}

\item{trainLabels}{are the labels for the training}

\item{trainLabels}{is the y vector}
}
\description{
This function is a very simple implementation of stochastic gradient descent.
Step size is set to 1/sqrt(t), which is a common stepsize, but one should be able to alter it.
}
\examples{


}
\keyword{Descent}
\keyword{Gradient}
\keyword{Stochastic}
