% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neural_network}
\alias{neural_network}
\title{Neural network (ANN)}
\usage{
neural_network(x, y, hidden_layers, activation_functions = "relu",
  task = "regression", weights_init = NULL)
}
\arguments{
\item{x}{are the features used in the neural network. Send in as a matrix or dataframe.}

\item{y}{are the response variables.}

\item{hidden_layers}{are the hidden layers in the network. Send in a a vector consisting of}

\item{activation_functions}{are sent in as a vector of strings, with as many activation functions as layers specified plus 1 for the final layer. If only one activation function is specified, this will be used throughout the whole network. Default is rectified linear units.}

\item{weights_init}{vector with initial weights.}

\item{verbose}{boolean should be outputted while training. Default set to false.}

\item{threshold}{is the threshold of the error when to stop training.}
}
\description{
Creates a fully connected ANN which can be trained will be trained with backpropagation.
Not done implemented. Remains TBD.
}
\keyword{ANN,}
\keyword{artificial,}
\keyword{backpropagation}
\keyword{backpropagation,}
\keyword{network,}
\keyword{neural,}
