% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMfunctions.R
\name{PerceptronSGD}
\alias{PerceptronSGD}
\title{Perceptron Stochastic Gradient Descent}
\usage{
PerceptronSGD(wInitVals, data, trainLabels, stepSizeFunc = cDivT, ...)
}
\arguments{
\item{wInitVals}{are the initial values of your vector}

\item{data}{is the features, i.e. the X vector.}

\item{trainLabels}{is the y vector}

\item{stepSizeFunc}{is the step size function used.}

\item{...}{is used to be able to specify additional parameters sent in to stepSizeFunc.}
}
\description{
This function is a very simple implementation of gradient descent in a perceptron.
Step size is set to 1/t, which is a common stepsize, but one should be able to alter it.
}
\examples{

trainLabels = y
data = X
w = PerceptronSGD(rnorm(nrow(data), mean = 0, sd = 1), data, trainLabels = y)

}
\keyword{Descent}
\keyword{Gradient}
\keyword{Stochastic}
