% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_imputation_effect}
\alias{plot_imputation_effect}
\title{Plot imputation effect}
\usage{
plot_imputation_effect(feature, impute_func = median, ...)
}
\arguments{
\item{feature}{is the feature to impute for.}

\item{impute_func}{is the function to use for imputation. Takes in, by default, all the non-NA feature values.}

\item{...}{are additional values to \code{impute_func}.}
}
\description{
Plots a what-if scenario and compares the scenarios what the distribution would look like if
}
