% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{PolynomialKernel}
\alias{PolynomialKernel}
\title{Polynomial Kernel}
\usage{
PolynomialKernel(x1, x2, degree)
}
\arguments{
\item{x1}{is the first vector sent in.}

\item{x2}{is the second vector sent in.}

\item{degree}{is the degree to which the}
}
\description{
Returns (1 + x1*x2)^degree for features of any dimension
}
\examples{


}
\keyword{Kernel,}
\keyword{Polynomial}
