% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_inference.R
\name{posterior_mean}
\alias{posterior_mean}
\title{Posterior mean}
\usage{
posterior_mean(samples, exclude_zeros = TRUE)
}
\arguments{
\item{excludes_zeros}{is a boolean specifying whether the zeros should be excluded for the inference. This should be true (which is default) when using bsts.}
}
\description{
Returns the posterior mean estimate, given a vector of samples
}
