% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{predict.GaussianNB}
\alias{predict.GaussianNB}
\title{Predict GaussianNB}
\usage{
\method{predict}{GaussianNB}(modelObj, X, ret_probabilities = F)
}
\arguments{
\item{modelObj}{is the model itself, the \code{GaussianNaiveBayes} model.}

\item{X}{is the matrix/dataframe of features.}

\item{ret_probabilities}{is a boolean deciding whether to return probabilities or class predictions. Default \code{FALSE}.}
}
\description{
Predicts for a Gaussian Naive Bayes model.
}
