% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{predict.LinearRegression}
\alias{predict.LinearRegression}
\title{Predict linear regression}
\usage{
\method{predict}{LinearRegression}(modelObject, X)
}
\arguments{
\item{modelObject}{is the LinearRegression model object with which to predict.}

\item{X}{is the data with which to predict. Ensure it is in the same order as the original matrix.}
}
\description{
Predicts using the model created in function \code{linear_reg}.
}
