% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{predict.MultinomialNB}
\alias{predict.MultinomialNB}
\title{Predict MultinomialNB}
\usage{
\method{predict}{MultinomialNB}(modelObject, X, ret_probabilities = F)
}
\arguments{
\item{modelObject}{is the multinomial naive bayes with which to predict.}

\item{X}{is the new data for which to predict. Should have same format as original data frame}

\item{ret_probabilities}{is a boolean deciding whether to return probabilities or class predictions. Default \code{FALSE}.}
}
\description{
Predictor for multinomial naive bayes.
}
