% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NaiveBayes.R
\name{predict.NB}
\alias{predict.NB}
\title{Predict Naive Bayes}
\usage{
\method{predict}{NB}(model, X, ret_probabilities = F)
}
\arguments{
\item{model}{is the model in question.}

\item{X}{is the matrix of features to predict for.}

\item{ret_probabilities}{is a boolean deciding whether to return probabilities or class predictions. Default \code{FALSE}.}
}
\description{
Prediction function for the \code{NB} class.
}
