% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_vals}
\alias{replace_vals}
\title{Replacer}
\usage{
replace_vals(column, ranking, corresponding, NA_val = NA)
}
\arguments{
\item{column}{is the vector to replace values for.}

\item{corresponding}{is the vector with values to replace.}

\item{NA_val}{is the value to replace \code{NA}-values with. Default is \code{NA}, meaning they are not replaced.}
}
\description{
Replaces values in a a vector \code{column} with values in a corresponding vector \code{ranking}.
}
