% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rescaleFeature}
\alias{rescaleFeature}
\title{rescaleFeature}
\usage{
rescaleFeature(feature, mu, stdDev)
}
\arguments{
\item{feature}{Vector of scaled features to rescale}

\item{mu}{to rescale the vector with.}

\item{stdDev}{to rescale the vector with.}
}
\description{
Standardizes a vector of features with mu and sd. Not sure yet if correct, not validated yet.
}
\examples{



}
\keyword{standardization}
