% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{resid.LinearRegression}
\alias{resid.LinearRegression}
\title{Residuals linear regression}
\usage{
\method{resid}{LinearRegression}(modelObject)
}
\arguments{
\item{modelObject}{is the LinearRegression model in question}
}
\description{
Returns the residuals of the fitted training data.
}
