% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{RMSLE}
\alias{RMSLE}
\title{Root Mean-squared logarithmic error}
\usage{
RMSLE(y, y_hat, logged = FALSE)
}
\arguments{
\item{y}{are the true value}

\item{y_hat}{are the predictions}

\item{logged}{is whether the vector are already logged or not. Default \code{FALSE}.}
}
\description{
Root Mean-squared logarithmic error
}
