% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardizeData}
\alias{standardizeData}
\title{Standardize data}
\usage{
standardizeData(dataIn, typeOut = "dataframe")
}
\arguments{
\item{dataIn}{is a matrix or a dataframe sent in to be standardized.}

\item{typeOut}{specifies whether we want a dataframe or matrix out}
}
\description{
Standardize all data in a data frame or a matrix that are of type double. Assumes normal distribution on data.
}
\examples{



}
\keyword{cluster}
