% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardizeFeature}
\alias{standardizeFeature}
\title{standardizeFeatures}
\usage{
standardizeFeature(feature, mu = NULL, stdDev = NULL)
}
\arguments{
\item{feature}{Vector of features to standardize according to its length}

\item{mu}{to standardize the vector with. If nothing is sent in, the mean of the feature vector will be used.}

\item{stdDev}{to standardize the vector with. If nothing is sent in, the sd of the feature vector will be used.}
}
\description{
Standardizes a vector of features with mu and sd. Not sure yet if correct, not validated yet.
}
\examples{



}
\keyword{standardization}
