% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{substitute_values}
\alias{substitute_values}
\title{Substitute values}
\usage{
substitute_values(column, ranking, corresponding, null_val = NULL)
}
\arguments{
\item{column}{is the vector of values from which the values should be replaced.}

\item{ranking}{are the values that are to be replaced in column. Should contain all possible values in the matrix.}

\item{corresponding}{are the corresponding values that replaces the values given in ranking. Order is important.}

\item{null_val}{is a value given as to what replace null values with.}
}
\description{
Substitutes values in a column with other given values.
}
\keyword{data,}
\keyword{manipulation}
