% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{summary.LinearRegression}
\alias{summary.LinearRegression}
\title{Summary linear regression}
\usage{
\method{summary}{LinearRegression}(modelObject)
}
\arguments{
\item{modelObject}{is the modelObject for which to return the summary.}
}
\description{
Summary linear regression
}
