% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVMfunctions.R
\name{testSVMAccuracy}
\alias{testSVMAccuracy}
\title{Kernelized Perceptron}
\usage{
testSVMAccuracy(w, test, Y_Test)
}
\arguments{
\item{w}{is the gradient vector}

\item{test}{is a frame or matrix with the features.}

\item{Y_Test}{are the labels corresponding.}
}
\description{
This function is a very simple implementation of a kernelized perceptron
Step size is set to 1/t, which is a common stepsize, but one should be able to alter it. This is TBD
Currently WAY too slow. No point in even using unless improvements are made.
}
\examples{





}
\keyword{SVM}
