% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_series_cv}
\alias{time_series_cv}
\title{Time-series cross-validation}
\usage{
time_series_cv(model, data, formula_ = NULL, target = NULL,
  predictors = NULL, start_index = 3, error_measure_function, ...)
}
\arguments{
\item{model}{is the model on which to perform the time-series cross-validation.}

\item{data}{is the data on which to perform the cross-validation and provide good measures for.}

\item{formula_}{is the formula to pass to the model function. \code{NULL} by default, and if it is \code{NULL}, \code{target} and \code{predictors} must be passed.}

\item{target}{is the name of the target variable. Can be \code{NULL} if \code{formula_} is specified.}

\item{predictors}{is a vector of the names of the predictors. Can be \code{NULL} if \code{formula_} is specified.}

\item{start_index}{is the first index to perform the CV from. Default is 3 (not reasonable to start from 2 and forward).}

\item{error_measure_function}{is the error function with which one measures the generalization error.}

\item{...}{are additional model parameters.}
}
\description{
Performs time-series cross-validation on a model.
}
\details{
Not tested yet.
}
