% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_regression.R
\name{to_polynomial}
\alias{to_polynomial}
\title{Get polynomial data}
\usage{
to_polynomial(X, degree = 2)
}
\arguments{
\item{X}{is the original dataframe.}

\item{degree}{is how many degrees to convert the frame to.}
}
\description{
Get a dataframe with all the original data frame's variables also as polynomial. Thus allows for polynomial regression.
}
\details{
Can be used in conjunction with \code{linear_reg} by using this function as \code{transform_func}.
}
