% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{TPR}
\alias{TPR}
\title{True positive rate}
\usage{
TPR(predictions, labels)
}
\arguments{
\item{predictions}{is the vector of predictions, with true labels sent in as 1, false as 0}

\item{labels}{is the vector of actual labels, with true labels sent in as 1, false as 0}
}
\description{
Calculates the true positive rate, given predictions and labels. Send in labels and predictions with 1 as true, 0 as false.
}
\examples{



}
\keyword{TPR}
