% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{trainTestSplit}
\usage{
trainTestSplit(dataSet, labels, testSetFraction, seed = NULL)
}
\arguments{
\item{dataSet}{is the set of data to split up. Assumes data set does not contain labels but is split up separately.}

\item{labels}{is a vector containing the labels for the dataSet. Needs to have the same order and length as dataSet (of course).}
}
\description{
Splits the data into a training and test set
}
\examples{



}
\keyword{Cross}
\keyword{validation}
