% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KNearest.R
\name{volume_N_sphere}
\alias{volume_N_sphere}
\title{Volume of an n-dimensional sphere}
\usage{
volume_N_sphere(nDim, radius)
}
\arguments{
\item{nDim}{is the number of dimensions}

\item{radius}{is the radius of the sphere, given in an absolute value.}
}
\description{
Returns the volume of an n-dimensional sphere
}
\examples{


}
\keyword{Volume,}
\keyword{sphere}
