% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R
\name{wMAPE}
\alias{wMAPE}
\title{wMAPE}
\usage{
wMAPE(y, y_hat)
}
\arguments{
\item{y}{is a vector of the actual values.}

\item{y_hat}{is a vector of the predictions.}
}
\description{
Calculates the wMAPE value for predictions. The wMAPE is calculated as \deqn{(\Sigma |y[i] - y_hat[i]|)/(\Sigma y[i])}
}
\details{
NOT tested yet!
}
