% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dummy_vars}
\alias{create_dummy_vars}
\title{Create dummy variables}
\usage{
create_dummy_vars(data, col_name)
}
\arguments{
\item{data}{is the data frame (send in the whole data frame)}

\item{col_name}{is the name of the column, given as a string. It can also be given as an index.}
}
\description{
Creates dummy variables for a specific column in a data frame.
}
\keyword{data,}
\keyword{dummy,}
\keyword{manipulation}
