% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FPR}
\alias{FPR}
\title{True positive rate}
\usage{
FPR(predictions, labels)
}
\arguments{
\item{predictions}{is the vector of predictions, with true labels sent in as 1, false as 0}

\item{labels}{is the vector of actual labels, with true labels sent in as 1, false as 0}
}
\description{
Calculates the false positive rate for predictions derived from a probability threshold, given predictions and labels. Send in labels and predictions in factor form.
}
\examples{



}
\keyword{FPR}
