% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{GaussianKernel}
\alias{GaussianKernel}
\title{Gaussian Kernel}
\usage{
GaussianKernel(x1, x2, h, sigmaF = 1)
}
\arguments{
\item{x1}{is the first datapoint.}

\item{x2}{is the second datapoint.}

\item{h}{is the variance used in the operation.}

\item{sigmaF}{is the \eqn{\sigma} before.}
}
\description{
Returns a gaussian kernel operation
}
\examples{


}
\keyword{Gaussian}
\keyword{Kernel,}
