% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{GaussianKernel}
\alias{GaussianKernel}
\title{Polynomial Kernel}
\usage{
GaussianKernel(x1, x2, h)
}
\arguments{
\item{x1}{is the first vector sent in.}

\item{x2}{is the second vector sent in.}

\item{h}{is the variance used in the operation.}
}
\description{
Returns a gaussian kernel operation
}
\examples{

 
}
\keyword{Gaussian}
\keyword{Kernel,}
