% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_measures.R, R/utils.R
\name{MSE}
\alias{MSE}
\title{MSE}
\usage{
MSE(y, y_hat)

MSE(y, y_hat)
}
\arguments{
\item{y}{is a vector containing the actual response values.}

\item{y_hat}{is a vector containing the predictions.}
}
\description{
Returns the MSE, given true and predicted values.

Calculates the mean squared error for predictions.
}
\details{
NOT tested yet!
}
\examples{



}
\keyword{error}
