% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kernels.R
\name{P_G_Kernel}
\alias{P_G_Kernel}
\title{Periodic + Gaussian Kernel}
\usage{
P_G_Kernel(sigmaF = 1, x1, x2, l1, l2, d)
}
\arguments{
\item{sigmaF}{is the constant before.}

\item{x1}{is the first datapoint.}

\item{x2}{is the second datapoint.}

\item{l1}{is the lengthscale for the periodic kernel.}

\item{l2}{is the lengthscale for the gaussian kernel.}

\item{d}{is for the periodic kernel.}
}
\description{
Returns a mix of a periodic and a gaussian kernel. This kernel will be \deqn{(sigmaF^2)*exp(-2*((sin(pi*abs(x1-x2)/d)^2)/(l1^2)))*exp(-0.5*((x1 - x2)^2)/(l2^2))}
}
\keyword{Gaussian}
\keyword{Kernel,}
